requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
backoff==1.11.1
charset-normalizer==2.0.9
distro==1.6.0
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
python-dateutil==2.8.2
ruyaml==0.91.0
six==1.16.0
typing_extensions==4.0.1
yarl==1.7.2
